Xcode 11.2.1 headers and strings

+++ Headers/DVTFoundation/DVTConsoleLogger.h
@@ -15,7 +15,7 @@
 @property unsigned long long indentLevel; // @synthesize indentLevel=_indentLevel;
 - (void)decreaseIndentLevelForAspect:(id)arg1;
 - (void)increaseIndentLevelForAspect:(id)arg1;
-- (void)logMessage:(id)arg1 forAspect:(id)arg2 logLevel:(int)arg3;
+- (void)logMessage:(id)arg1 forAspect:(id)arg2 logLevel:(unsigned long long)arg3;
 
 @end
