+++ Headers/DVTFoundation/DVTLogger.h
@@ -16,7 +16,7 @@
 + (void)initialize;
 - (void)decreaseIndentLevelForAspect:(id)arg1;
 - (void)increaseIndentLevelForAspect:(id)arg1;
-- (void)logMessage:(id)arg1 forAspect:(id)arg2 logLevel:(int)arg3;
+- (void)logMessage:(id)arg1 forAspect:(id)arg2 logLevel:(unsigned long long)arg3;
 - (id)stringForIndentLevel:(unsigned long long)arg1;
 - (id)threadPrefixString;
