+++ Headers/DVTFoundation/DVTMemorySnapshot.h
@@ -19,8 +19,8 @@
 + (id)snapshotForLogAspect:(id)arg1;
 + (void)initialize;
 - (void).cxx_destruct;
-- (void)takeVmmapSnapshotWithLogLevel:(int)arg1 logFormat:(id)arg2;
-- (void)_takeVmmapSnapshotWithLogLevel:(int)arg1 toFilePath:(id)arg2 logString:(id)arg3;
+- (void)takeVmmapSnapshotWithLogLevel:(unsigned long long)arg1 logFormat:(id)arg2;
+- (void)_takeVmmapSnapshotWithLogLevel:(unsigned long long)arg1 toFilePath:(id)arg2 logString:(id)arg3;
 - (unsigned long long)_newIncrementForSnapshotType:(id)arg1;
 - (id)initWithAspect:(id)arg1;
