+++ Headers/DVTKit/DVTTextlikeFindDescriptor-Protocol.h
@@ -7,6 +7,6 @@
 #import <DVTKit/NSObject-Protocol.h>
 
 @protocol DVTTextlikeFindDescriptor <NSObject>
-@property(readonly) int matchStyle;
+@property(readonly) unsigned long long matchStyle;
 @end
