+++ Headers/IDEFoundation/IDEDistributionFileLogger.h
@@ -8,12 +8,12 @@
 
 @interface IDEDistributionFileLogger : DVTFileLogger
 {
-    int _exclusiveLogLevel;
+    unsigned long long _exclusiveLogLevel;
 }
 
-@property(readonly) int exclusiveLogLevel; // @synthesize exclusiveLogLevel=_exclusiveLogLevel;
-- (void)logMessage:(id)arg1 forAspect:(id)arg2 logLevel:(int)arg3;
-- (id)initWithURL:(id)arg1 append:(BOOL)arg2 exclusiveLogLevel:(int)arg3;
+@property(readonly) unsigned long long exclusiveLogLevel; // @synthesize exclusiveLogLevel=_exclusiveLogLevel;
+- (void)logMessage:(id)arg1 forAspect:(id)arg2 logLevel:(unsigned long long)arg3;
+- (id)initWithURL:(id)arg1 append:(BOOL)arg2 exclusiveLogLevel:(unsigned long long)arg3;
 
 @end
