+++ Headers/IDEInterfaceBuilderCocoaIntegration/IDEInterfaceBuilderAutomator.h
@@ -16,7 +16,7 @@
     NSImage *_automatorIcon;
 }
 
-+ (BOOL)ide_initializeWithOptions:(int)arg1 error:(id *)arg2;
++ (BOOL)ide_initializeWithOptions:(unsigned long long)arg1 error:(id *)arg2;
 + (id)sharedInstance;
 - (void).cxx_destruct;
 - (void)setupAutomator;
