+++ Headers/IDEInterfaceBuilderCocoaIntegration/IDEInterfaceBuilderCocoa.h
@@ -12,7 +12,7 @@
 {
 }
 
-+ (BOOL)ide_initializeWithOptions:(int)arg1 error:(id *)arg2;
++ (BOOL)ide_initializeWithOptions:(unsigned long long)arg1 error:(id *)arg2;
 + (id)sharedInstance;
 - (void)setupPDFKit;
 - (void)setupImageKit;
