+++ Headers/IDEKit/IDEDeferredInitialization-Protocol.h
@@ -5,6 +5,6 @@
 //
 
 @protocol IDEDeferredInitialization
-+ (BOOL)ide_deferredInitializeWithOptions:(int)arg1 error:(id *)arg2;
++ (BOOL)ide_deferredInitializeWithOptions:(unsigned long long)arg1 error:(id *)arg2;
 @end
