+++ Headers/IDEKit/IDESourceControlConflictResolutionController.h
@@ -15,7 +15,7 @@
 }
 
 + (void)replaceItemAtFilePath:(id)arg1 withFileWrapper:(id)arg2;
-+ (int)binaryConflictMergeDirection:(id)arg1;
++ (long long)binaryConflictMergeDirection:(id)arg1;
 + (BOOL)isBinaryConflictResolutionMergeData:(id)arg1;
 @property BOOL shouldDisableMergeSaving; // @synthesize shouldDisableMergeSaving=_shouldDisableMergeSaving;
 - (void).cxx_destruct;
